lab/requirements-dev.txt
seleniumbase
mypy
django-stubs
djangorestframework-stubs
